pyproject.toml
[project]
name = "agent-vis"
version = "0.1.0"
description = "Workflow visualization with real-time flow animations"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "websockets>=12.0",
    "requests>=2.31.0",
    "websocket-client>=1.7.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-playwright>=0.5.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "wsproto>=1.2.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.7.0",
    "pylint>=3.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_vis"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

